Added a proposal for a BIAB setup with false bottom and electrical heater
This is an initial proposal of a BIAB setup that would allow to automate
almost each steps of the brewing process.
There are 2 water inputs: clean water and well water
During boiling process, we need to add the cooler coil.
The setup is made for 30L kettle, but this can be scaled for 10L or 50L.

Signed-off-by: Jean-Patrice Laude <jpatrice.laude@gmail.com>